% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_complete_fn.R
\name{keep_complete}
\alias{keep_complete}
\title{The Keep Complete function}
\usage{
keep_complete(df)
}
\arguments{
\item{df}{The dataframe you want to clean!}
}
\description{
This function takes a dataset with NA values and returns that dataset cleaned of all rows that contain an NA value.
}
\examples{
keep_complete()
mtcars[mtcars>160] <- NA
mtcars = keep_complete(mtcars)
print(mtcars)
}
\keyword{cleaning,}
\keyword{complete.cases,}
\keyword{data}
\keyword{tidying}
