% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMinionOut.R
\name{parseMinionOut}
\alias{parseMinionOut}
\title{Parse the output from Minion}
\usage{
parseMinionOut(out)
}
\arguments{
\item{out}{Output from Minion as a string.}
}
\value{
List of data.frames containing the parsed results.
}
\description{
Parse the output from Minion
}
